+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## 0.13.2
+
+Released 2024-09-24
+
+### Fixes
+
+* [enforce capacity](https://github.com/ParkenDD/parkapi-sources-v3/pull/133)
+* [remove confusing herrenberg field](https://github.com/ParkenDD/parkapi-sources-v3/pull/134)
+* [set kienzler public url](https://github.com/ParkenDD/parkapi-sources-v3/pull/135)
+* [fix Herrenberg parking type mapping](https://github.com/ParkenDD/parkapi-sources-v3/pull/136)
+
+
+## 0.13.1
+
+Release 2024-09-22
+
+### Fixes
+
+* [Add VRS data](https://github.com/ParkenDD/parkapi-sources-v3/pull/123)
+* [Fix bahn mapping](https://github.com/ParkenDD/parkapi-sources-v3/pull/124)
+* [Fix PBW Mapping](https://github.com/ParkenDD/parkapi-sources-v3/pull/125)
+* [Fix Heidelberg fee](https://github.com/ParkenDD/parkapi-sources-v3/pull/126)
+* [Split up kienzler requests](https://github.com/ParkenDD/parkapi-sources-v3/pull/128)
+* [Fix bfrk mapping](https://github.com/ParkenDD/parkapi-sources-v3/pull/129)
+
+
+## 0.13.0
+
+Released 2024-09-16
+
+### Features
+
+* [Herrenberg static bike pull converter](https://github.com/ParkenDD/parkapi-sources-v3/pull/110)
+
+
+### Fixes
+
+* [Park and Ride at BFRK](https://github.com/ParkenDD/parkapi-sources-v3/pull/121)
+
+
 ## 0.12.1
 
 Released 2024-09-09